requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
opencv-python>=4.5.5
matplotlib>=3.5.1
scikit-learn>=1.0.2
pandas>=1.3.5
seaborn>=0.11.2
imbalanced-learn>=0.9.0
fastapi>=0.75.0
uvicorn>=0.17.6
python-multipart>=0.0.5
python-dotenv>=0.20.0
streamlit>=1.9.0
Pillow>=9.0.1
albumentations>=1.1.0
